Use cp instead of rsync + systemd-tmpfiles for cleaning
This use cp with hardlinks + a bind mount for keeping the removed kernel
modules accessible.
Cleaning up the backup is done by systemd-tmpfiles on next boot.
Cleaning up the mount is either done on shutdown implicitly, or when we
reinstall the running kernel by the pre-hook script.

No systemd service enabling is needed.